+++ Kernel/System/Notification.pm
@@ -104,8 +104,8 @@ sub NotificationGet {
         );
 
         # convert subject
-        $Data[3] = $EncodeObject->Convert(
-            Text  => $Data[3],
+        $Data[4] = $EncodeObject->Convert(
+            Text  => $Data[4],
             From  => $Data[2],
             To    => 'utf-8',
             Force => 1,